requirements.txt
pytesseract
Pillow
PyMuPDF